expand project cmd for more model/dataset details
