+++ client/starwhale/cli/cluster.py
@@ -1,6 +1,6 @@
 import click
 from rich import print as rprint
-from starwhale.cluster import Cluster
+from starwhale.cluster import ClusterView
 
 
 @click.group("cluster", help="starwhale cluster info and terminal ui")
@@ -10,7 +10,7 @@ def cluster_cmd():
 
 @cluster_cmd.command("info", help="show controller cluster info")
 def _info():
-    Cluster().info()
+    ClusterView().info()
 
 
 @cluster_cmd.command("tui", help="basic terminal ui for starwhale controller")
