+++ client/starwhale/cli/project.py
@@ -1,6 +1,6 @@
 import click
 
-from starwhale.cluster import Cluster, DEFAULT_PAGE_NUM, DEFAULT_PAGE_SIZE
+from starwhale.cluster import ClusterView, DEFAULT_PAGE_NUM, DEFAULT_PAGE_SIZE
 
 
 @click.group("project", help="starwhale controller project info and operation")
@@ -12,8 +12,9 @@ def project_cmd():
 @click.option("-a", "--all-users", is_flag=True, help="list all users project, if not set, cli will show current user's projects")
 @click.option("-p", "--page", default=DEFAULT_PAGE_NUM, help="page number for projects list")
 @click.option("-s", "--size", default=DEFAULT_PAGE_SIZE, help="page size for projects list")
-def _list(all_users, page, size):
-    Cluster().list(all_users, page, size)
+@click.option("--fullname", is_flag=True, help="show version fullname")
+def _list(all_users, page, size, fullname):
+    ClusterView().list(all_users, page, size, fullname)
 
 
 @project_cmd.command("create", help="create a new project in starwhale controller")
